requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
redis[hiredis]==5.2.1
asyncpg==0.30.0
PyNaCl==1.5.0
base58==2.1.1
